requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
defusedxml==0.7.1
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
drf-nested-routers==0.94.2
httpie==3.2.4
idna==3.10
Markdown==3.8
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.4
mysqlclient==2.2.7
pillow==11.2.1
psycopg2==2.9.10
pycparser==2.22
Pygments==2.19.1
PyMySQL==1.1.1
PySocks==1.7.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
setuptools==80.9.0
sqlparse==0.5.3
urllib3==2.4.0
